move util.hash to dhis2
