+++ packages/common/src/dateFns.js
@@ -12,16 +12,4 @@ export {
    * @function
    */
   format,
-  /**
-   * The isValid function from the date-fns library. See {@link https://date-fns.org/v3.6.0/docs/isValid}
-   * @public
-   * @function
-   */
-  isValid,
-  /**
-   * The addDays function from the date-fns library. See {@link https://date-fns.org/v3.6.0/docs/addDays}
-   * @public
-   * @function
-   */
-  addDays,
 } from 'date-fns';
