+++ packages/common/src/util/index.js
@@ -4,5 +4,4 @@ export * from './references.js';
 export { default as parseDate } from './parse-date.js';
 export { default as throwError } from './throw-error.js';
 export { encode, decode } from './base64.js';
-export { hash, sha256 } from './hash.js';
 export { uuid } from './uuid.js';
